Broke down compose items and removed dependant state.
