+++ wear/src/main/java/io/homeassistant/companion/android/home/views/LandingScreen.kt
@@ -1,0 +1,203 @@
+package io.homeassistant.companion.android.home.views
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import androidx.wear.compose.material.Chip
+import androidx.wear.compose.material.PositionIndicator
+import androidx.wear.compose.material.Scaffold
+import androidx.wear.compose.material.ScalingLazyColumn
+import androidx.wear.compose.material.ScalingLazyListState
+import androidx.wear.compose.material.Text
+import androidx.wear.compose.material.rememberScalingLazyListState
+import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.data.integration.Entity
+import io.homeassistant.companion.android.home.HomeActivity
+import io.homeassistant.companion.android.home.HomePresenterImpl
+import io.homeassistant.companion.android.util.RotaryEventDispatcher
+import io.homeassistant.companion.android.util.RotaryEventState
+import io.homeassistant.companion.android.util.SetTitle
+import io.homeassistant.companion.android.util.setChipDefaults
+
+@Composable
+fun LandingScreen(
+    entities: MutableList<Entity<Any>>,
+    favorites: MutableSet<String>,
+    onEntityClicked: (String) -> Unit,
+    onSettingsClicked: () -> Unit,
+    onLogoutClicked: () -> Unit
+) {
+    val scalingLazyListState: ScalingLazyListState = rememberScalingLazyListState()
+
+    var expandedFavorites: Boolean by rememberSaveable { mutableStateOf(true) }
+    var expandedInputBooleans: Boolean by rememberSaveable { mutableStateOf(true) }
+    var expandedLights: Boolean by rememberSaveable { mutableStateOf(true) }
+    var expandedScenes: Boolean by rememberSaveable { mutableStateOf(true) }
+    var expandedScripts: Boolean by rememberSaveable { mutableStateOf(true) }
+    var expandedSwitches: Boolean by rememberSaveable { mutableStateOf(true) }
+
+    val scenes =
+        entities.sortedBy { it.entityId }.filter { it.entityId.split(".")[0] == "scene" }
+    val scripts =
+        entities.sortedBy { it.entityId }.filter { it.entityId.split(".")[0] == "script" }
+    val lights =
+        entities.sortedBy { it.entityId }.filter { it.entityId.split(".")[0] == "light" }
+    val inputBooleans =
+        entities.sortedBy { it.entityId }.filter { it.entityId.split(".")[0] == "input_boolean" }
+    val switches =
+        entities.sortedBy { it.entityId }.filter { it.entityId.split(".")[0] == "switch" }
+
+    RotaryEventDispatcher(scalingLazyListState)
+    RotaryEventState(scrollState = scalingLazyListState)
+
+    Scaffold(
+        positionIndicator = {
+            if (scalingLazyListState.isScrollInProgress)
+                PositionIndicator(scalingLazyListState = scalingLazyListState)
+        }
+    ) {
+        ScalingLazyColumn(
+            modifier = Modifier
+                .fillMaxSize(),
+            contentPadding = PaddingValues(
+                top = 10.dp,
+                start = 10.dp,
+                end = 10.dp,
+                bottom = 40.dp
+            ),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            state = scalingLazyListState
+        ) {
+            if (favorites.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.favorites,
+                        expanded = expandedFavorites,
+                        onExpandChanged = { expandedFavorites = it }
+                    )
+                }
+                val favoriteArray = favorites.toTypedArray()
+                if (expandedFavorites) {
+                    items(favoriteArray.size) { index ->
+                        FavoriteEntities(
+                            favoriteArray[index],
+                            null,
+                            onEntityClicked
+                        )
+                    }
+                }
+            }
+            if (entities.isNullOrEmpty()) {
+                item {
+                    Column {
+                        SetTitle(id = R.string.loading)
+                        Chip(
+                            modifier = Modifier
+                                .padding(
+                                    top = 10.dp,
+                                    start = 10.dp,
+                                    end = 10.dp
+                                ),
+                            label = {
+                                Text(
+                                    text = stringResource(R.string.loading_entities),
+                                    textAlign = TextAlign.Center
+                                )
+                            },
+                            onClick = { /* No op */ },
+                            colors = setChipDefaults()
+                        )
+                    }
+                }
+            }
+            if (inputBooleans.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.input_booleans,
+                        expanded = expandedInputBooleans,
+                        onExpandChanged = { expandedInputBooleans = it }
+                    )
+                }
+                if (expandedInputBooleans) {
+                    items(inputBooleans.size) { index ->
+                        EntityUi(inputBooleans[index], onEntityClicked)
+                    }
+                }
+            }
+            if (lights.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.lights,
+                        expanded = expandedLights,
+                        onExpandChanged = { expandedLights = it }
+                    )
+                }
+                if (expandedLights) {
+                    items(lights.size) { index ->
+                        EntityUi(lights[index], onEntityClicked)
+                    }
+                }
+            }
+            if (scenes.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.scenes,
+                        expanded = expandedScenes,
+                        onExpandChanged = { expandedScenes = it }
+                    )
+                }
+                if (expandedScenes) {
+                    items(scenes.size) { index ->
+                        EntityUi(scenes[index], onEntityClicked)
+                    }
+                }
+            }
+            if (scripts.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.scripts,
+                        expanded = expandedScripts,
+                        onExpandChanged = { expandedScripts = it }
+                    )
+                }
+                if (expandedScripts) {
+                    items(scripts.size) { index ->
+                        EntityUi(scripts[index], onEntityClicked)
+                    }
+                }
+            }
+            if (switches.isNotEmpty()) {
+                item {
+                    SetListHeader(
+                        stringId = R.string.switches,
+                        expanded = expandedSwitches,
+                        onExpandChanged = { expandedSwitches = it }
+                    )
+                }
+                if (expandedSwitches) {
+                    items(switches.size) { index ->
+                        EntityUi(switches[index], onEntityClicked)
+                    }
+                }
+            }
+            item {
+                LoadOtherSection(
+                    onSettingsClicked = onSettingsClicked,
+                    onLogoutClicked = onLogoutClicked
+                )
+            }
+        }
+    }
+}
